refactor: Update imports in session and init files; remove obsolete test file
